// src/shared/interface/themeColors.interface.ts
export interface ThemeColors {
  background: string;
  splashbackground: string;
  text: string;
  buttonBackground: string;
  red: string;
  reverseText: string;
  inputBackground: string;
  inputBorder: string;
  eyeColor: string;
  forgotPasswordColor: string;
  black: string;
  dontHaveAnAccountColor: string;
  inputPlaceholderColor: string;
  signupGradient: string;
  signupGradient2: string;
  numberPlateText: string;
  numberPlateStateColor: string;
  tabBarIconActive: string;
  tabBarIconInactive: string;
  tabBarbackground: string;
  white: string;
  bannerHeading: string;
  bannerSubheading: string;
  backButtonColor: string;
  backButtonIconColor: string;
  profileTab:string;
  profileImageBackGround:string;
  settingTabsSeparator:string;
  myWinningsCard:string;
  myWinningsCardlabel:string;
  myWinningsCardValue:string;
  settingTabsIcons:string;
  settingTabsTitles:string;
  cameraIconContainer:string;
  languageModalTabBackground:string;
  languageModalTabBorder:string;
  languageModalDoneColor:string;
  giftsBackground:string;
  categoryChipBackground:string;
  selectedGiftCheckBoxContainer:string;
  submitModalDetails:string;
  submitModalOverlay:string;
  monthlyAccessHeading:string;
  monthlyAccessDetail:string;
  gradientTextSecond:string;
  whiteOpacity6:string;
  whiteOpacity7:string;
  proSubacriptionGardient:string;
  imageBackgroundColor:string;
  actionSheetOverlay:string;
  pricingCardBoder:string;
  pricingCardSubtitle:string;
  shimmerMainColor:string;
  shimmerSecondaryColor:string;
  toastBorder:string;
  toastBackground:string;


  

}